+++ src/main/java/lotto/view/ConsoleInputReader.java
@@ -52,7 +52,7 @@ public class ConsoleInputReader {
         System.out.println("수동으로 구매할 로또 수를 입력해 주세요.");
 
         try {
-            final long lottoCount = Long.parseLong(scanner.nextLine());
+            final int lottoCount = Integer.parseInt(scanner.nextLine());
 
             return new LottoCount(lottoCount);
         } catch (NumberFormatException e) {
